Consolidate documentation into module.md files (#142)
* Remove examples.md in favor of module.md

* Move all docs to modules

* Post merge problems.

* Do not use gh specific markdown.

* Adding the dht docs.

* Replace DHT for one that's well known.
