+++ dids/module.md
@@ -1,34 +1,22 @@
 # Module dids
 
-# Package web5.sdk.dids.methods.key
+This module contains the code for generating and resolving DIDs of different major methods.
+
+The documentations of each package contains examples on how to use the SDK. It's meant as a quick reference on how to
+effectively use this SDK.
+
+Since documentation is always a living document, there's a chance that some examples might be outdated.
+If you find yourself scratching your head because something doesn't work, the best place
+to look at is the test files. We strive to maintain a high test coverage, which should give
+you plenty of examples on how to use this SDK.
+
+# Package web5.sdk.dids.methods.dht
+
+Package that contains the `DidDht` class, which is used to create and resolve dids using the `dht` method.
 
 # Examples
 
-## Create & Resolve
-
-### Creating a DID Key
-
-```kt
-package example
-
-import web5.sdk.crypto.InMemoryKeyManager
-import web5.sdk.dids.methods.key.DidKey
-
-val keyManager = InMemoryKeyManager()
-val did = DidKey.create(keyManager)
-```
-
-### Creating a DID Jwk
-
-```kt
-package example
-
-import web5.sdk.crypto.InMemoryKeyManager
-import web5.sdk.dids.methods.jwk.DidJwk
-
-val keyManager = InMemoryKeyManager()
-val did = DidJwk.create(keyManager)
-```
+## Creation
 
 ### Creating a DID DHT
 
@@ -58,15 +46,137 @@ val opts = CreateDidDhtOptions(
 val did = DidDht.create(keyManager, opts)
 ```
 
-### Resolving a DID
+## Resolution
+
+### Resolve a DID DHT
+
+```kotlin
+val did = DidDht.resolve("did:dht:gb46emk73wkenrut43ii67a3o5qctojcaucebth7r83pst6yeh8o")
+```
+
+# Package web5.sdk.dids.methods.ion
+
+Package that contains the `DidIon` class, which is used to create and resolve dids using the `ion` method.
+
+# Examples
+
+## Creation
+
+### Create an ION did
+
+This is the simplest way to create an ion did.
+
+```kotlin
+val did = DidIon.create(InMemoryKeyManager())
+```
+
+The private keys will be stored in the `InMemoryKeyManager`. All the defaults are used for
+the `DidIonApi`, including the endpoint for the ION node used for creation, and uses
+`CIO` as the `HttpClientEngine` (see [ktor engines](https://ktor.io/docs/http-client-engines.html)).
+
+### Create an ION did with custom ION endpoint and engine
+
+```kotlin
+val keyManager = InMemoryKeyManager()
+val ionApi = DidIonApi {
+  ionHost = "my_custom_ion_host"
+  engine = CIO.create {
+    maxConnectionsCount = 10
+    requestTimeout = 5.toDuration(DurationUnit.SECONDS).inWholeMilliseconds
+  }
+}
+val did = ionApi.create(keyManager)
+```
+
+### Create an ION did with custom creation options
+
+This is considered an advanced use case.
+
+Make sure that you have access to the private keys associated with the public keys you're passing in
+(i.e. `verification`, `update`, and `recovery` keys). You can generate with the `keyManager`, or
+store them elsewhere however you see fit.
+
+```kotlin
+val keyManager = InMemoryKeyManager()
+val opts = CreateDidIonOptions(
+  verificationMethodsToAdd = listOf(
+    VerificationMethodCreationParams(
+      JWSAlgorithm.ES256K,
+      relationships = listOf(PublicKeyPurpose.AUTHENTICATION, PublicKeyPurpose.ASSERTION_METHOD)
+    ),
+    VerificationMethodCreationParams(
+      JWSAlgorithm.ES256K,
+      relationships = listOf(PublicKeyPurpose.ASSERTION_METHOD)
+    ),
+  )
+  val did = DidIon . create (keyManager, opts)
+```
+
+## Resolution
+
+### Resolve an ION did
+
+```kotlin
+val didResolutionResult = DidIon.resolve("did:ion:EiClkZMDxPKqC9c-umQfTkR8vvZ9JPhl_xLDI9Nfk38w5w")
+```
+
+## DID ION Operations
+
+### Recover an ION did
+
+Any sidetree based DID, including ION, supports
+a [recover operation](https://identity.foundation/sidetree/spec/#recover).
+This type of operation is useful when the update keys of your DID have been compromised.
+
+```kotlin
+// We create the DID first. 
+val keyManager = InMemoryKeyManager()
+val did = DidIon.create(keyManager)
+val recoveryKeyAlias = did.creationMetadata!!.keyAliases.verificationKeyAlias
+
+// Imagine that your update key was compromised, so you need to recover your DID.
+val opts = RecoverDidIonOptions(
+  recoveryKeyAlias = recoveryKeyAlias.first(),
+)
+val recoverResult = DidIon.recover(keyManager, did.uri, opts)
+```
+
+**NOTE**: The `keyManager` MUST contain the recovery private key.
+
+### Deactivate an ION did
+
+```kotlin
+// We create the DID first. 
+val ionApi = DidIon
+val keyManager = InMemoryKeyManager()
+val did = ionApi.create(keyManager)
+val recoveryKeyAlias = did.creationMetadata!!.keyAliases.verificationKeyAlias.first()
+
+// You want to permanently disable the DID, rendering it useless.
+val opts = DeactivateDidIonOptions(
+  recoveryKeyAlias = recoveryKeyAlias,
+)
+val deactivateResult = ionApi.deactivate(keyManager, did.uri, opts)
+```
+
+**NOTE**: The `keyManager` MUST contain the recovery private key.
+
+# Package web5.sdk.dids.methods.key
+
+Package that contains the `DidKey` class, which is used to create and resolve dids using the `key` method.
+
+## Examples
+
+### Create a DID Key
 
 ```kt
 package example
 
-import web5.sdk.dids.DidResolvers
+import web5.sdk.crypto.InMemoryKeyManager
+import web5.sdk.dids.methods.key.DidKey
 
-// did is the same as the value from the example above
-val didResolutionResult = DidResolvers.resolve(did.uri)
+val keyManager = InMemoryKeyManager()
+val did = DidKey.create(keyManager)
 ```
 
 ## Export / Import
@@ -131,3 +241,33 @@ fun main() {
 ```
 
 json serializing or deserializing an exported keyset is _not_ limited to Jackson. It can be done using any json library.
+
+# Package web5.sdk.dids.methods.web
+
+Package that contains the `DidWeb` class, which is used to create and resolve dids using the `web` method.
+
+## Examples
+
+### Resolve a Web did
+
+```kotlin
+val didResolutionResult = DidWeb.resolve("did:web:users.tbddev.org:demo")
+```
+
+# Package web5.sdk.dids.methods.jwk
+
+Package that contains the `DidJwk` class, which is used to create and resolve dids using the `jwk` method.
+
+## Examples
+
+### Creating a DID Jwk
+
+```kt
+package example
+
+import web5.sdk.crypto.InMemoryKeyManager
+import web5.sdk.dids.methods.jwk.DidJwk
+
+val keyManager = InMemoryKeyManager()
+val did = DidJwk.create(keyManager)
+```
