+++ dids/src/main/kotlin/web5/sdk/dids/methods/web/DidWeb.kt
@@ -1,4 +1,4 @@
-package web5.sdk.dids
+package web5.sdk.dids.methods.web
 
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import foundation.identity.did.DID
@@ -15,7 +15,13 @@ import io.ktor.http.isSuccess
 import io.ktor.serialization.jackson.jackson
 import kotlinx.coroutines.runBlocking
 import web5.sdk.crypto.KeyManager
+import web5.sdk.dids.CreateDidOptions
+import web5.sdk.dids.Did
+import web5.sdk.dids.DidMethod
+import web5.sdk.dids.DidResolutionResult
+import web5.sdk.dids.ResolveDidOptions
 import web5.sdk.dids.methods.ion.InvalidStatusException
+import web5.sdk.dids.validateKeyMaterialInsideKeyManager
 import java.net.URL
 import java.net.URLDecoder
 import kotlin.text.Charsets.UTF_8
